projects: ad469x_evb: use the example infrastructure
Migrate ad469x_evb to the shared examples infrastructure (examples.mk):
replace the Makefile BASIC_EXAMPLE/IIO_EXAMPLE toggles with EXAMPLE := iio
+ include examples.mk, give the iio example an IIOD=y example.mk (its IIO
device is defined inline, so no extra driver source), drop examples_src.mk
and platform_includes.h, and shorten the example directories
(basic_example -> basic, iio_example -> iio). The project-wide
AD469X_DEV/AD469X_TEMP/AD469X_SEQ selects stay in src.mk. The xilinx
main.c keeps its cache setup and defers to example_main(); the iio example
now includes iio_app.h directly (it used to come in transitively via
platform_includes.h). Switch builds.json flags to EXAMPLE= (keeping the
AD469X_SEQ variants and hardware arrays) and update the README's build
commands to EXAMPLE=<name>.

Verified by dependency resolution (no local xilinx toolchain): both
examples resolve their full SRCS/INCS with the correct device and
sequencer defines, and every referenced file exists.
