+++ projects/ad469x_evb/builds.json
@@ -1,19 +1,19 @@
 {
       "xilinx": {
         "basic": {
-              "flags" : "IIO_EXAMPLE=n BASIC_EXAMPLE=y",
+              "flags" : "EXAMPLE=basic",
               "hardware" : ["ad469x_evb_zed"]
         },
         "iio": {
-              "flags" : "IIO_EXAMPLE=y BASIC_EXAMPLE=n",
+              "flags" : "EXAMPLE=iio",
               "hardware" : ["ad469x_evb_zed"]
         },
         "standard_seq": {
-              "flags" : "IIO_EXAMPLE=n BASIC_EXAMPLE=y AD469X_SEQ=standard",
+              "flags" : "EXAMPLE=basic AD469X_SEQ=standard",
               "hardware" : ["ad469x_evb_zed"]
         },
         "advance_seq": {
-              "flags" : "IIO_EXAMPLE=n BASIC_EXAMPLE=y AD469X_SEQ=advanced",
+              "flags" : "EXAMPLE=basic AD469X_SEQ=advanced",
               "hardware" : ["ad469x_evb_zed"]
         }
       }
