+++ projects/ad469x_evb/src.mk
@@ -1,16 +1,3 @@
-include $(PROJECT)/src/platform/$(PLATFORM)/platform_src.mk
-include $(PROJECT)/src/examples/examples_src.mk
-
-SRCS += $(PROJECT)/src/platform/$(PLATFORM)/main.c
-
-INCS += $(PROJECT)/src/common/common_data.h
-SRCS += $(PROJECT)/src/common/common_data.c
-
-INCS += $(PROJECT)/src/platform/platform_includes.h
-
-INCS += $(PROJECT)/src/platform/$(PLATFORM)/parameters.h
-SRCS += $(PROJECT)/src/platform/$(PLATFORM)/parameters.c
-
 SRCS += $(DRIVERS)/api/no_os_uart.c		\
 	$(DRIVERS)/api/no_os_gpio.c		\
 	$(NO-OS)/util/no_os_list.c		\
