+++ projects/ad469x_evb/src/common/common_data.h
@@ -33,7 +33,7 @@
 #ifndef __COMMON_DATA_H__
 #define __COMMON_DATA_H__
 
-#include "platform_includes.h"
+#include "parameters.h"
 #include "ad469x.h"
 
 extern struct no_os_uart_init_param iio_uart_ip;
