--- projects/ad469x_evb/src/examples/examples_src.mk
@@ -1,17 +1,0 @@
-ifeq (y,$(strip $(IIO_EXAMPLE)))
-CFLAGS += -DIIO_EXAMPLE
-SRCS += $(PROJECT)/src/examples/iio_example/iio_example.c
-INCS += $(PROJECT)/src/examples/iio_example/iio_example.h
-IIOD=y
-
-SRC_DIRS += $(NO-OS)/iio/iio_app
-
-INCS += $(INCLUDE)/no_os_list.h \
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.h
-endif
-
-ifeq (y,$(strip $(BASIC_EXAMPLE)))
-CFLAGS += -DBASIC_EXAMPLE
-SRCS += $(PROJECT)/src/examples/basic_example/basic_example.c
-INCS += $(PROJECT)/src/examples/basic_example/basic_example.h
-endif
