+++ projects/ad469x_evb/src/examples/iio/example.mk
@@ -1,0 +1,1 @@
+IIOD = y
