+++ projects/ad469x_evb/src/platform/xilinx/main.c
@@ -31,14 +31,11 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
 
-#include "platform_includes.h"
+#include "parameters.h"
 #include "common_data.h"
 #include "no_os_error.h"
-#ifdef IIO_EXAMPLE
-#include "iio_example.h"
-#elif defined(BASIC_EXAMPLE)
-#include "basic_example.h"
-#endif
+
+extern int example_main();
 
 /***************************************************************************//**
  * @brief Main function execution for xilinx platform.
@@ -47,18 +44,8 @@
 *******************************************************************************/
 int main()
 {
-	int ret = -EINVAL;
-
 	Xil_DCacheDisable();
 	Xil_ICacheDisable();
 
-#ifdef IIO_EXAMPLE
-	ret = iio_example_main();
-#elif defined (BASIC_EXAMPLE)
-	ret = basic_example_main();
-#else
-#error At least one example has to be selected using y value in Makefile.
-#endif
-
-	return ret;
+	return example_main();
 }
